refactor(SPV-1021): change helpers package to response, removed unnecesary pointer
